Use XdndProxy to receive drag and drop events in VST windows
This gets drag and drop working on hosts that support such on Linux,
specifically works in:

- Tracktion
- Bitwig

Does not work in:

- Ardour
- Reaper

The basic idea of the patch is that it has to set a proxy on the parent
window so that the Sitala window receives Xdnd events.  After setting
that, we then need the JUCE side of things so that it correctly maps
proxied calls to the correct peer.